refactor(datastorage): split server.go into focused modules
Refactored the 1,046-line server.go file into smaller, focused modules:

**New Structure:**
- pkg/datastorage/adapter/db_adapter.go (~300 lines) - Core DB operations
- pkg/datastorage/adapter/aggregations.go (~312 lines) - Aggregation queries
- pkg/datastorage/adapter/utils.go (~50 lines) - Utility functions
- pkg/datastorage/server/server.go (~354 lines) - HTTP server & lifecycle
- pkg/datastorage/server/handlers.go (~132 lines) - HTTP handlers & middleware
- pkg/datastorage/server/config.go (~28 lines) - Configuration struct

**Benefits:**
- Smaller, focused files (<400 lines each)
- Clear separation of concerns (adapter vs server)
- Easier navigation and maintenance
- Better testability

**Test Results:**
- Unit tests: 475/475 passed
- Integration tests: 123/123 passed

**Test Fix:**
- Fixed test isolation issue in aggregation_api_test.go by adding cleanup
  in BeforeAll to remove leftover data from previous test suites
